+++ docs/html/search/classes_0.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['charwriter_110',['CharWriter',['../classace__segment_1_1CharWriter.html',1,'ace_segment']]],
-  ['clockinterface_111',['ClockInterface',['../classace__segment_1_1ClockInterface.html',1,'ace_segment']]],
-  ['clockwriter_112',['ClockWriter',['../classace__segment_1_1ClockWriter.html',1,'ace_segment']]]
+  ['charwriter_96',['CharWriter',['../classace__segment_1_1CharWriter.html',1,'ace_segment']]],
+  ['clockinterface_97',['ClockInterface',['../classace__segment_1_1ClockInterface.html',1,'ace_segment']]],
+  ['clockwriter_98',['ClockWriter',['../classace__segment_1_1ClockWriter.html',1,'ace_segment']]]
 ];