+++ docs/html/search/functions_6.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['hc595module_169',['Hc595Module',['../classace__segment_1_1Hc595Module.html#ae3ce4d5d32c92c01df7a424c6e5e56cc',1,'ace_segment::Hc595Module']]],
-  ['ht16k33module_170',['Ht16k33Module',['../classace__segment_1_1Ht16k33Module.html#ad5a97827d1c639656c5685b61140f977',1,'ace_segment::Ht16k33Module']]]
+  ['hc595module_147',['Hc595Module',['../classace__segment_1_1Hc595Module.html#ae3ce4d5d32c92c01df7a424c6e5e56cc',1,'ace_segment::Hc595Module']]],
+  ['ht16k33module_148',['Ht16k33Module',['../classace__segment_1_1Ht16k33Module.html#a0136851a409d3bba51c3cdf9aa4eb483',1,'ace_segment::Ht16k33Module']]]
 ];