+++ docs/html/search/functions_a.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['numberwriter_180',['NumberWriter',['../classace__segment_1_1NumberWriter.html#abb61529de2cbf876061915895d68f01d',1,'ace_segment::NumberWriter']]]
+  ['numberwriter_160',['NumberWriter',['../classace__segment_1_1NumberWriter.html#abb61529de2cbf876061915895d68f01d',1,'ace_segment::NumberWriter']]]
 ];