+++ docs/html/search/functions_e.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['temperaturewriter_201',['TemperatureWriter',['../classace__segment_1_1TemperatureWriter.html#a3d0dc943b6b9dbb9ab43ac1720189bd1',1,'ace_segment::TemperatureWriter']]],
-  ['tm1637module_202',['Tm1637Module',['../classace__segment_1_1Tm1637Module.html#a9cc401ceeed8292b26a93d5dcc08312e',1,'ace_segment::Tm1637Module']]]
+  ['temperaturewriter_174',['TemperatureWriter',['../classace__segment_1_1TemperatureWriter.html#a3d0dc943b6b9dbb9ab43ac1720189bd1',1,'ace_segment::TemperatureWriter']]],
+  ['tm1637module_175',['Tm1637Module',['../classace__segment_1_1Tm1637Module.html#a9cc401ceeed8292b26a93d5dcc08312e',1,'ace_segment::Tm1637Module']]]
 ];