+++ examples/AutoBenchmark/generate_readme.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python3
+#!/usr/bin/env python3
 #
 # Python script that regenerates the README.md from the embedded template. Uses
 # ./generate_table.awk to regenerate the ASCII tables from the various *.txt
@@ -20,7 +20,6 @@ esp32_results = check_output(
     "./generate_table.awk < esp32.txt", shell=True, text=True)
 teensy32_results = check_output(
     "./generate_table.awk < teensy32.txt", shell=True, text=True)
-#teensy32_results = 'TBD'
 
 print(f"""\
 # AutoBenchmark
@@ -29,15 +28,18 @@ This program creates instances of various subclasses `LedModule` using different
 configurations:
 
 * `DirectModule`: group and segment pins directly connected to MCU
-* `DirectFast4Module`: same as `DirectModule` but using `digitalWriteFast` library
+* `DirectFast4Module`: same as `DirectModule` but using `digitalWriteFast`
+  library
 * `HybridModule`: group pins connected directly to MCU, but segment pins
-  connected to 74HC595 accessed through SPI
-* `Hc595Module`: group pins and segment pins connected to two 74HC595
+  connected to one 74HC595 that is accessed through SPI
+* `Hc595Module`: group pins and segment pins connected to two 74HC595 chips
   which are accessed through SPI
-* `Tm1637Module`: an LED module using a TM1637 driver chip, accessed through
-  a two-wire protocol similar to I2C
-* `Max7219Module`: an LED module using a MAX7219 driver chip, accessed through
+* `Tm1637Module`: an LED module using a TM1637 controller, accessed through
+  a custom two-wire protocol similar to I2C
+* `Max7219Module`: an LED module using a MAX7219 controller, accessed through
   SPI
+* `Ht16k33Module`: an LED module using an HT16K33 controller, accessed through
+  I2C
 
 We then measure the time taken by the methods used to render the digit to the
 LED module:
@@ -50,8 +52,10 @@ LED module:
     * Sends out the buffered digits over the custom two-wire protocol.
 * `Max7219Module::flush()`
     * Sends out the buffered digits using SPI.
+* `Ht16k33Module::flush()`
+    * Sends out the buffered digits using I2C
 
-**Version**: AceSegment v0.7
+**Version**: AceSegment v0.8
 
 **DO NOT EDIT**: This file was auto-generated using `make README.md`.
 
@@ -62,10 +66,11 @@ This program depends on the following libraries:
 * [AceCommon](https://github.com/bxparks/AceCommon)
 * [AceSegment](https://github.com/bxparks/AceButton)
 
-On AVR processors, the following library is required to run the
+On AVR processors, one of the following libraries is required to run the
 `digitalWriteFast()` versions of the low-level drivers:
 
-* [digitalWriteFast](https://github.com/NicksonYap/digitalWriteFast)
+* https://github.com/watterott/Arduino-Libs/tree/master/digitalWriteFast
+* https://github.com/NicksonYap/digitalWriteFast
 
 ## How to Generate
 
@@ -147,10 +152,11 @@ number of `TimingStats::update()` calls that were made.
   actually sustain a 20 MHz SPI, which breaks the MAX7219 chip because it can
   handle only 16 MHz.
 * Verified that removing the 10 nF capacitors from the CLK and DIO lines of the
-  TM1637 modules allows a much shorter `BIT_DELAY`. Added CPU benchmark numbers
-  for `BIT_DELAY = 5` microseconds (e.g. `Tm1637(4,SoftTmi,5us)` and
-  `Tm1637(4,SoftTmiFast,5us)`). The `flush()` or `flushIncremental()` durations
-  are almost a factor of 10X to 20X shorter compared to `BIT_DELAY = 100`.
+  TM1637 modules allows a much shorter `DELAY_MICROS`. Added CPU benchmark
+  numbers for `DELAY_MICROS = 5` microseconds (e.g. `Tm1637(4,SimpleTmi,5us)`
+  and `Tm1637(4,SimpleTmiFast,5us)`). The `flush()` or `flushIncremental()`
+  durations are almost a factor of 10X to 20X shorter compared to `DELAY_MICROS
+  = 100`.
 
 **v0.7:**
 
@@ -171,41 +177,101 @@ number of `TimingStats::update()` calls that were made.
           dev boards that run AutoBenchmark. Prevents clock stretching even if
           no LED module is actually attached to the pins.
 
+**v0.8**
+
+* Extract communication interfaces into AceSPI, AceTMI, and AceWire libraries.
+  No change.
+* Copy interface objects into various modules (Hc595Module, Max7219Module,
+  Tm1637Module, Ht16k33Module) by value instead of by reference.
+    * Remove an extra layer of indirection.
+    * Makes almost no difference in execution speed. Maybe if I squint hard
+      enough, it looks like a few microseconds faster on average?
+    * Saves flash consumption on AVR processors (see MemoryBenchmark/README.md).
+* Regenerate after attaching an actual HT16K33 LED module on the I2C bus.
+    * The hardware I2C library `<Wire.h>` checks for a proper ACK
+      response from the I2C device. If a NACK is received, then it stops
+      transmitting the rest of the data in the transmit buffer and returns
+      immediately from `endTransmission()`.
+    * Causes AutoBenchmark to produces incorrect timing (too short) for the
+      `Ht16k33(TwoWire)` benchmark.
+    * Software I2C implementations such as `SimpleWireInterface` and
+      `SimpleWireFastInterface` do *not* check the response from the I2C device.
+      The benchmark results are not affected.
+    * For AVR processors, the software I2C implementation of
+      `SimpleWireFastInterface` using `digitalWriteFast()` is *faster* than
+      hardware I2C using `<Wire.h>`:
+        * `<Wire.h>` with `setClock(400000)` produces throughput of about 234
+          kHz.
+        * `SimpleWireFastInterface` using 1us delayMicros produces a throughput
+          of about 350 kHz.
+    * For 32-bit processors without a `digitalWriteFast` library,
+      `SimpleWireFastInterface` is comparable to the 100 kHz setting of
+      `<Wire.h>`, until we get to very fast 32-bit processors like the ESP32 and
+      Teensy 3.2, where `SimpleWireInterface` becomes competitive with 400 kHz
+      of `<Wire.h>`.
+
 ## Results
 
 The following tables show the number of microseconds taken by:
 
 * `DirectModule::renderFieldNow()`, `HybridModule::renderFieldNow()`,
   `Hc595Module::renderFieldNow()`
-    * renders the 8 segments of a single LED digit.
+    * Renders a single LED digit.
     * If the LED module has 4 digits, then `renderFieldNow()` must be called 4
-      times to render the light pattern of the entire LED module. The entire
-      rendering is then called a frame.
+      times to render the light pattern of the entire LED module. Each digit
+      is called a "field" and entire rendering of 4 fields is then called a
+      "frame".
     * Most people can no longer see flickering of the display at about 60 frames
       a second. To achieve that, the `renderFieldNow()` method must be called
       240 times a second for a module with 4 digits, or every 4.17 milliseconds.
     * The results below show that every processor, even the slowest AVR
       processor, is able to meet this threshhold.
+    * Several communication implemenations are tested:
+        * SimpleSpi: software bitbanging SPI using `digitalWrite()`
+        * SimpleSpiFast: software bitbanging SPI using `<digitalWriteFast.h>`
+        * HardSpi: hardware SPI using `<SPI.h>` and `digitalWrite()` to control
+          the latch pin
+        * HardSpiFast: hardware SPI using `<SPI.h>` and `<digitalWriteFast.h>``
+        * to control the latch pin
+    * The `subfield` label indicates that `NUM_SUBFIELDS = 16` was used to
+      obtain PWM brightness modulation on a per-digit basis.
+        * Each call to `renderFieldNow()` performs a rendering of one of the 16
+          subfields of a given digit, which gives us the PWM.
+        * The average duration goes down, but the maximum remains the same.
 * `Tm1637Module::flush()` or `Tm1637Module::flushIncremental()`
-    * sends digits in the buffer to the TM1637 LED module using the I2C-like
-      protocol
-    * results for two values of `BIT_DELAY` are collected
-        * 100 microseconds (e.g. `Tm1637(4,SoftTmi)`)
-        * 5 microseconds (e.g. `Tm1637(4,SoftTmi,5us)`)
+    * Sends digits in the buffer to the TM1637 LED module using the I2C-like
+      protocol.
+    * Results for two values of `DELAY_MICROS` are collected:
+        * SimpleTmi: software bigbanging the TM1637 protocol using
+          `digitalWrite()` with a `delayMicros` of 100  microseconds or 5
+          microseconds
+        * SimpleTmiFast: software bigbanging the TM1637 protocol using
+          `<digitalWriteFast.h>` and a `DELAY_MICROS` of 100  microseconds or 5
+          microseconds
 * `Max7219Module::flush()`
-    * sends all digits in the buffer to the MAX7219 LED module using
-      software SPI or hardware SPI
+    * Sends all digits in the buffer to the MAX7219 LED module using 4 types of
+      communcation interfaces
+        * SimpleSpi: software SPI using `digitalWrite()`
+        * SimpleSpiFast: software SPI using `<digitalWriteFast.h>`
+        * HardSpi: hardware SPI using `<SPI.h>` and `digitalWrite()` to control
+          the latch pin
+        * HardSpiFast: hardware SPI using `<SPI.h>` and `<digitalWriteFast.h>`
+          to control the latch pin
 * `Ht16k33Module::flush()`
-    * sends all digits in the buffer to the HT16K33 LED module using
-      hardware I2C
+    * Sends all 4 digits in the buffer and the brightness setting to the HT16K33
+      LED module using 2 communication interfaces:
+        * TwoWire: hardware I2C using `<Wire.h>`
+        * SimpleWire: AceSegment's custom bitbanging I2C implementation
+        * SimpleWireFast: Same as SimpleWire using `<digitalWriteFast.h>`
+    * Total bits: the addr, 5 x 16-bit words, then another addr, and the
+      brightness, for a total of 13 bytes. Each byte sends 8 bits and reads the
+      ACK bit from the slave, so 9 bits per byte. Total bits: 117 bits.
 
-On AVR processors, the "fast" options are available using the
-[digitalWriteFast](https://github.com/NicksonYap/digitalWriteFast) library whose
-`digitalWriteFast()` functions can be up to 50X faster if the `pin` number and
-`value` parameters are compile-time constants. In addition, the
-`digitalWriteFast` functions reduce flash memory consumption by 600-700 bytes
-for `SoftTmiFastInterface`, `SoftSpiFastInterface`, and `HardSpiFastInterface`
-compared to their non-fast equivalents.
+On AVR processors, the "fast" options are available using one of the
+digitalWriteFast libraries whose `digitalWriteFast()` functions can be up to 50X
+faster if the `pin` number and `value` parameters are compile-time constants. In
+addition, the `digitalWriteFast` functions reduce flash memory consumption by
+600-700 bytes compared to their non-fast equivalents.
 
 ### Arduino Nano